pyproject.toml
[project]
name = "stravaboard"
version = "1.0.0"
description = "A dashboard for flexibly displaying and tracking Strava runs."
authors = [
    { name = "David Zhang" },
]
requires-python = ">=3.13"
dependencies = [
    "requests==2.32.5",
    "pandas==2.3.2",
    "plotly==6.3.0",
    "python-dotenv==1.1.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit==4.3.0",
    "pytest==8.4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["pip>=22.0", "setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
